update components and removed `NumberText`
