+++ screens/AppNavigator/screens/BalancesScreen/SendScreen/SendScreen.tsx
@@ -7,8 +7,9 @@ import { useForm, Controller, Control } from 'react-hook-form'
 import { BarCodeScanner } from 'expo-barcode-scanner'
 import NumberFormat from 'react-number-format'
 import tailwind from 'tailwind-rn'
-import { PrimaryButton, Text, TextInput } from '../../../../../components'
+import { Text, TextInput } from '../../../../../components'
 import { getTokenIcon } from '../../../../../components/icons/tokens/_index'
+import { PrimaryButton } from '../../../../../components/PrimaryButton'
 import { PrimaryColor, PrimaryColorStyle } from '../../../../../constants/Theme'
 import { translate } from '../../../../../translations'
 import { BalanceParamList } from '../BalancesNavigator'
